+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.3.0] - 2024-10-18
+
 ### Added
 
 - ✨(api) add RELEASE version on config endpoint #459
@@ -102,7 +104,8 @@ and this project adheres to
 - ✨(domains) create and manage domains on admin + API
 - ✨(domains) mailbox creation + link to email provisioning API
 
-[unreleased]: https://github.com/numerique-gouv/people/compare/v1.2.1...main
+[unreleased]: https://github.com/numerique-gouv/people/compare/v1.3.0...main
+[1.3.0]: https://github.com/numerique-gouv/people/releases/v1.3.0
 [1.2.1]: https://github.com/numerique-gouv/people/releases/v1.2.1
 [1.2.0]: https://github.com/numerique-gouv/people/releases/v1.2.0
 [1.1.0]: https://github.com/numerique-gouv/people/releases/v1.1.0